Refactor security pool validation and fork/auction test workflows
- update mainnet deployment addresses in docs to latest deterministic outputs
- extract shared truth-auction setup logic into reusable test helpers
- refactor SecurityPool and SecurityPoolOracleCoordinator to centralize checks and clearer staged-operation error handling
- add UI fork-auction/workflow interaction state hooks and improve security pool orchestration test coverage
