+++ solidity/ts/tests/securityPoolForkerStorageLayout.test.ts
@@ -2,9 +2,13 @@ import { test } from 'bun:test'
 import assert from '../testsuite/simulator/utils/assert'
 import { getArray, getContractOutput, getRecord, getString, loadContractsJson, normalizeStorageLayout } from './contractArtifactHelpers'
 
+function getForkerStorageLayout(artifacts: Record<string, unknown>, sourcePath: string, contractName: string) {
+	return normalizeStorageLayout(getContractOutput(artifacts, sourcePath, contractName))
+}
+
 test('SecurityPoolForker retains unified own-fork fields in fork session storage', () => {
 	const artifacts = loadContractsJson(import.meta.dir)
-	const forkerLayout = normalizeStorageLayout(getContractOutput(artifacts, 'contracts/peripherals/SecurityPoolForker.sol', 'SecurityPoolForker'))
+	const forkerLayout = getForkerStorageLayout(artifacts, 'contracts/peripherals/SecurityPoolForker.sol', 'SecurityPoolForker')
 	const forkDataByPoolEntry = forkerLayout.find(entry => entry.label === 'forkDataByPool')
 	if (forkDataByPoolEntry === undefined) throw new Error('Storage layout missing forkDataByPool field')
 	const forkDataByPoolValueType = getRecord(forkDataByPoolEntry.type.value, 'Storage layout missing forkDataByPool value type')
@@ -13,4 +17,14 @@ test('SecurityPoolForker retains unified own-fork fields in fork session storage
 	assert.ok(forkDataLabels.has('vaultRepAtFork'))
 	assert.ok(forkDataLabels.has('escalationChildRepAtFork'))
 	assert.ok(forkDataLabels.has('escalationSourceRepAtFork'))
+})
+
+test('SecurityPoolForker delegates keep the exact host storage layout', () => {
+	const artifacts = loadContractsJson(import.meta.dir)
+	const hostLayout = getForkerStorageLayout(artifacts, 'contracts/peripherals/SecurityPoolForker.sol', 'SecurityPoolForker')
+	const vaultMigrationDelegateLayout = getForkerStorageLayout(artifacts, 'contracts/peripherals/SecurityPoolForkerVaultMigrationDelegate.sol', 'SecurityPoolForkerVaultMigrationDelegate')
+	const escalationGameForkerLayout = getForkerStorageLayout(artifacts, 'contracts/peripherals/EscalationGameForker.sol', 'EscalationGameForker')
+
+	assert.deepStrictEqual(vaultMigrationDelegateLayout, hostLayout)
+	assert.deepStrictEqual(escalationGameForkerLayout, hostLayout)
 })
