+++ ui/ts/contracts/securityPools.ts
@@ -1,4 +1,4 @@
-import { decodeEventLog, encodeAbiParameters, encodeDeployData, getCreate2Address, keccak256, parseAbiItem, zeroAddress, type Address, type TransactionReceipt } from 'viem'
+import { decodeEventLog, encodeAbiParameters, encodeDeployData, getCreate2Address, keccak256, parseAbiItem, zeroAddress, type Address, type ContractFunctionParameters, type TransactionReceipt } from 'viem'
 import {
 	peripherals_EscalationGame_EscalationGame,
 	peripherals_SecurityPoolOracleCoordinator_SecurityPoolOracleCoordinator,
@@ -21,7 +21,14 @@ import { loadMarketDetails } from './zoltar.js'
 
 const QUESTION_OUTCOME_ABI = [parseAbiItem('function getQuestionOutcome(address securityPool) view returns (uint8 outcome)')]
 
-const ACTIVE_SECURITY_POOL_VAULT_PREVIEW_LIMIT = 3n
+const SECURITY_POOL_LIST_VAULT_PREVIEW_LIMIT = 50n
+const SECURITY_POOL_PAGE_VAULT_PREVIEW_LIMIT = 3n
+
+export type LoadAllSecurityPoolsOptions = {
+	accountAddress?: Address
+	selectedSecurityPoolAddress?: Address | string
+	vaultDetailMode?: 'all' | 'selected'
+}
 
 type SecurityPoolDeploymentQueryResult = {
 	parent: Address
@@ -139,7 +146,7 @@ async function loadSecurityPoolVaultSummaries(
 	vaults: ListedSecurityPool['vaults']
 }> {
 	const vaultCount = await getSecurityPoolVaultCount(client, securityPoolAddress)
-	const previewLimit = options.previewLimit ?? ACTIVE_SECURITY_POOL_VAULT_PREVIEW_LIMIT
+	const previewLimit = options.previewLimit ?? SECURITY_POOL_LIST_VAULT_PREVIEW_LIMIT
 	const previewCount = vaultCount < previewLimit ? vaultCount : previewLimit
 	const previewVaultAddresses = previewCount === 0n ? [] : await getSecurityPoolVaults(client, securityPoolAddress, 0n, previewCount)
 	const summaryVaultAddresses = [...previewVaultAddresses]
@@ -153,18 +160,14 @@ async function loadSecurityPoolVaultSummaries(
 			vaults: [],
 		}
 	}
+	const securityVaultSummaryContracts: ContractFunctionParameters[] = summaryVaultAddresses.map(vaultAddress => ({
+		abi: peripherals_SecurityPool_SecurityPool.abi,
+		functionName: 'securityVaults',
+		address: securityPoolAddress,
+		args: [vaultAddress],
+	}))
 	const [vaultData, totalRepBalance, poolOwnershipDenominator, escrowedRepByVault] = await Promise.all([
-		Promise.all(
-			summaryVaultAddresses.map(
-				async vaultAddress =>
-					await client.readContract({
-						abi: peripherals_SecurityPool_SecurityPool.abi,
-						functionName: 'securityVaults',
-						address: securityPoolAddress,
-						args: [vaultAddress],
-					}),
-			),
-		).then(result => requireSecurityVaultTupleArray(result, 'security vault tuple')),
+		readRequiredMulticall(client, securityVaultSummaryContracts).then(result => requireSecurityVaultTupleArray(result, 'security vault tuple')),
 		client.readContract({
 			abi: peripherals_SecurityPool_SecurityPool.abi,
 			functionName: 'getTotalRepBalance',
@@ -200,6 +203,204 @@ async function loadSecurityPoolVaultSummaries(
 			]
 		}),
 	}
+}
+
+function shouldLoadSecurityPoolVaults(
+	deployment: Pick<SecurityPoolDeploymentQueryResult, 'parent' | 'securityPool'>,
+	options: {
+		selectedSecurityPoolAddress?: Address | string
+		vaultDetailMode: 'all' | 'selected'
+	},
+) {
+	if (options.vaultDetailMode === 'all') return true
+	if (options.selectedSecurityPoolAddress === undefined) return false
+	return sameAddress(deployment.securityPool, options.selectedSecurityPoolAddress) || sameAddress(deployment.parent, options.selectedSecurityPoolAddress)
+}
+
+function createDeferredSecurityPoolVaultSummary(vaultCount: bigint) {
+	return {
+		hasLoadedVaults: vaultCount === 0n,
+		vaultCount,
+		vaults: [] as ListedSecurityPool['vaults'],
+	}
+}
+
+async function loadSecurityPoolDetails(
+	client: ReadClient,
+	deployment: SecurityPoolDeploymentQueryResult,
+	options: {
+		accountAddress?: Address
+		selectedSecurityPoolAddress?: Address | string
+		vaultDetailMode: 'all' | 'selected'
+		vaultPreviewLimit: bigint
+	},
+): Promise<ListedSecurityPool> {
+	const { parent, priceOracleManagerAndOperatorQueuer: managerAddress, questionId, securityMultiplier, securityPool: securityPoolAddress, truthAuction: truthAuctionAddress, universeId } = deployment
+	const shouldLoadVaults = shouldLoadSecurityPoolVaults(deployment, options)
+	const [[completeSetCollateralAmount, currentRetentionRate, forkData, lastOraclePrice, lastSettlementTimestamp, questionOutcome, systemStateValue, shareTokenSupply, totalRepDeposit, totalSecurityBondAllowance, universeForkTime], marketDetails, vaultSummaries] = await Promise.all([
+		readRequiredMulticall(client, [
+			{
+				abi: peripherals_SecurityPool_SecurityPool.abi,
+				functionName: 'completeSetCollateralAmount',
+				address: securityPoolAddress,
+				args: [],
+			},
+			{
+				abi: peripherals_SecurityPool_SecurityPool.abi,
+				functionName: 'currentRetentionRate',
+				address: securityPoolAddress,
+				args: [],
+			},
+			{
+				abi: peripherals_SecurityPoolForker_SecurityPoolForker.abi,
+				functionName: 'forkData',
+				address: getInfraContractAddresses().securityPoolForker,
+				args: [securityPoolAddress],
+			},
+			{
+				abi: peripherals_SecurityPoolOracleCoordinator_SecurityPoolOracleCoordinator.abi,
+				functionName: 'lastPrice',
+				address: managerAddress,
+				args: [],
+			},
+			{
+				abi: peripherals_SecurityPoolOracleCoordinator_SecurityPoolOracleCoordinator.abi,
+				functionName: 'lastSettlementTimestamp',
+				address: managerAddress,
+				args: [],
+			},
+			{
+				abi: QUESTION_OUTCOME_ABI,
+				functionName: 'getQuestionOutcome',
+				address: getInfraContractAddresses().securityPoolForker,
+				args: [securityPoolAddress],
+			},
+			{
+				abi: peripherals_SecurityPool_SecurityPool.abi,
+				functionName: 'systemState',
+				address: securityPoolAddress,
+				args: [],
+			},
+			{
+				abi: peripherals_SecurityPool_SecurityPool.abi,
+				functionName: 'shareTokenSupply',
+				address: securityPoolAddress,
+				args: [],
+			},
+			{
+				abi: peripherals_SecurityPool_SecurityPool.abi,
+				functionName: 'getTotalRepBalance',
+				address: securityPoolAddress,
+				args: [],
+			},
+			{
+				abi: peripherals_SecurityPool_SecurityPool.abi,
+				functionName: 'totalSecurityBondAllowance',
+				address: securityPoolAddress,
+				args: [],
+			},
+			{
+				abi: Zoltar_Zoltar.abi,
+				functionName: 'getForkTime',
+				address: getInfraContractAddresses().zoltar,
+				args: [universeId],
+			},
+		]),
+		loadMarketDetails(client, questionId),
+		shouldLoadVaults
+			? loadSecurityPoolVaultSummaries(client, securityPoolAddress, {
+					...(options.accountAddress === undefined ? {} : { accountAddress: options.accountAddress }),
+					previewLimit: options.vaultPreviewLimit,
+				})
+			: getSecurityPoolVaultCount(client, securityPoolAddress).then(createDeferredSecurityPoolVaultSummary),
+	])
+	const { truthAuctionStartedAt, migratedRep, forkOwnSecurityPool, forkOutcomeIndex } = requireForkDataView(forkData)
+	const forkOutcome = getForkOutcomeKey(forkOutcomeIndex, parent)
+	const systemState = getSecurityPoolSystemState(systemStateValue)
+	return {
+		completeSetCollateralAmount,
+		currentRetentionRate,
+		forkOutcome,
+		forkOwnSecurityPool,
+		hasForkActivity: deriveHasForkActivity({
+			forkOutcome,
+			migratedRep,
+			systemState,
+			truthAuctionStartedAt,
+		}),
+		lastOraclePrice: lastSettlementTimestamp > 0n ? lastOraclePrice : undefined,
+		lastOracleSettlementTimestamp: lastSettlementTimestamp,
+		managerAddress,
+		marketDetails,
+		migratedRep,
+		parent,
+		questionOutcome: getReportingOutcomeKey(questionOutcome),
+		questionId: getQuestionIdHex(questionId),
+		securityMultiplier,
+		securityPoolAddress,
+		shareTokenSupply,
+		systemState,
+		totalRepDeposit,
+		totalSecurityBondAllowance,
+		truthAuctionAddress,
+		truthAuctionStartedAt,
+		universeHasForked: universeForkTime > 0n,
+		universeId,
+		hasLoadedVaults: vaultSummaries.hasLoadedVaults,
+		vaultCount: vaultSummaries.vaultCount,
+		vaults: vaultSummaries.vaults,
+	}
+}
+
+async function loadSecurityPoolDeployments(client: ReadClient, startIndex: bigint, count: bigint) {
+	if (count === 0n) return [] as readonly SecurityPoolDeploymentQueryResult[]
+	return (await client.readContract({
+		address: getInfraContractAddresses().securityPoolFactory,
+		abi: peripherals_factories_SecurityPoolFactory_SecurityPoolFactory.abi,
+		functionName: 'securityPoolDeploymentsRange',
+		args: [startIndex, count],
+	})) as readonly SecurityPoolDeploymentQueryResult[]
+}
+
+async function loadListedSecurityPools(
+	client: ReadClient,
+	deployments: readonly SecurityPoolDeploymentQueryResult[],
+	options: {
+		accountAddress?: Address
+		selectedSecurityPoolAddress?: Address | string
+		vaultDetailMode: 'all' | 'selected'
+		vaultPreviewLimit: bigint
+	},
+) {
+	return await Promise.all(deployments.map(async deployment => await loadSecurityPoolDetails(client, deployment, options)))
+}
+
+function applyChildForkActivityHints(pools: ListedSecurityPool[]) {
+	return pools.map(pool => {
+		if (pool.hasForkActivity) return pool
+		if (!pools.some(candidate => sameAddress(candidate.parent, pool.securityPoolAddress))) return pool
+		return {
+			...pool,
+			hasForkActivity: true,
+		}
+	})
+}
+
+export async function loadAllSecurityPools(client: ReadClient, options: LoadAllSecurityPoolsOptions = {}): Promise<ListedSecurityPool[]> {
+	const deploymentCount = await client.readContract({
+		address: getInfraContractAddresses().securityPoolFactory,
+		abi: peripherals_factories_SecurityPoolFactory_SecurityPoolFactory.abi,
+		functionName: 'securityPoolDeploymentCount',
+		args: [],
+	})
+	const deployments = await loadSecurityPoolDeployments(client, 0n, deploymentCount)
+	const pools = await loadListedSecurityPools(client, deployments, {
+		...(options.accountAddress === undefined ? {} : { accountAddress: options.accountAddress }),
+		...(options.selectedSecurityPoolAddress === undefined ? {} : { selectedSecurityPoolAddress: options.selectedSecurityPoolAddress }),
+		vaultDetailMode: options.vaultDetailMode ?? 'all',
+		vaultPreviewLimit: SECURITY_POOL_LIST_VAULT_PREVIEW_LIMIT,
+	})
+	return applyChildForkActivityHints(pools)
 }
 
 export async function createSecurityPool(
@@ -250,129 +451,12 @@ export async function loadSecurityPoolPage(client: ReadClient, pageIndex: number
 		}
 	}
 	const count = poolCount - startIndex < BigInt(pageSize) ? poolCount - startIndex : BigInt(pageSize)
-	const deployments: readonly SecurityPoolDeploymentQueryResult[] = await client.readContract({
-		address: getInfraContractAddresses().securityPoolFactory,
-		abi: peripherals_factories_SecurityPoolFactory_SecurityPoolFactory.abi,
-		functionName: 'securityPoolDeploymentsRange',
-		args: [startIndex, count],
+	const deployments = await loadSecurityPoolDeployments(client, startIndex, count)
+	const pools = await loadListedSecurityPools(client, deployments, {
+		...(accountAddress === undefined ? {} : { accountAddress }),
+		vaultDetailMode: 'all',
+		vaultPreviewLimit: SECURITY_POOL_PAGE_VAULT_PREVIEW_LIMIT,
 	})
-	const pools = await Promise.all(
-		deployments.map(async deployment => {
-			const { parent, priceOracleManagerAndOperatorQueuer: managerAddress, questionId, securityMultiplier, securityPool: securityPoolAddress, truthAuction: truthAuctionAddress, universeId } = deployment
-			const [[completeSetCollateralAmount, currentRetentionRate, forkData, lastOraclePrice, lastSettlementTimestamp, questionOutcome, systemStateValue, shareTokenSupply, totalRepDeposit, totalSecurityBondAllowance, universeForkTime], marketDetails, vaultSummaries] = await Promise.all([
-				readRequiredMulticall(client, [
-					{
-						abi: peripherals_SecurityPool_SecurityPool.abi,
-						functionName: 'completeSetCollateralAmount',
-						address: securityPoolAddress,
-						args: [],
-					},
-					{
-						abi: peripherals_SecurityPool_SecurityPool.abi,
-						functionName: 'currentRetentionRate',
-						address: securityPoolAddress,
-						args: [],
-					},
-					{
-						abi: peripherals_SecurityPoolForker_SecurityPoolForker.abi,
-						functionName: 'forkData',
-						address: getInfraContractAddresses().securityPoolForker,
-						args: [securityPoolAddress],
-					},
-					{
-						abi: peripherals_SecurityPoolOracleCoordinator_SecurityPoolOracleCoordinator.abi,
-						functionName: 'lastPrice',
-						address: managerAddress,
-						args: [],
-					},
-					{
-						abi: peripherals_SecurityPoolOracleCoordinator_SecurityPoolOracleCoordinator.abi,
-						functionName: 'lastSettlementTimestamp',
-						address: managerAddress,
-						args: [],
-					},
-					{
-						abi: QUESTION_OUTCOME_ABI,
-						functionName: 'getQuestionOutcome',
-						address: getInfraContractAddresses().securityPoolForker,
-						args: [securityPoolAddress],
-					},
-					{
-						abi: peripherals_SecurityPool_SecurityPool.abi,
-						functionName: 'systemState',
-						address: securityPoolAddress,
-						args: [],
-					},
-					{
-						abi: peripherals_SecurityPool_SecurityPool.abi,
-						functionName: 'shareTokenSupply',
-						address: securityPoolAddress,
-						args: [],
-					},
-					{
-						abi: peripherals_SecurityPool_SecurityPool.abi,
-						functionName: 'getTotalRepBalance',
-						address: securityPoolAddress,
-						args: [],
-					},
-					{
-						abi: peripherals_SecurityPool_SecurityPool.abi,
-						functionName: 'totalSecurityBondAllowance',
-						address: securityPoolAddress,
-						args: [],
-					},
-					{
-						abi: Zoltar_Zoltar.abi,
-						functionName: 'getForkTime',
-						address: getInfraContractAddresses().zoltar,
-						args: [universeId],
-					},
-				]),
-				loadMarketDetails(client, questionId),
-				loadSecurityPoolVaultSummaries(client, securityPoolAddress, {
-					...(accountAddress === undefined ? {} : { accountAddress }),
-					previewLimit: ACTIVE_SECURITY_POOL_VAULT_PREVIEW_LIMIT,
-				}),
-			])
-			const { truthAuctionStartedAt, migratedRep, forkOwnSecurityPool, forkOutcomeIndex } = requireForkDataView(forkData)
-			const forkOutcome = getForkOutcomeKey(forkOutcomeIndex, parent)
-			const systemState = getSecurityPoolSystemState(systemStateValue)
-			const hasForkActivity = deriveHasForkActivity({
-				forkOutcome,
-				migratedRep,
-				systemState,
-				truthAuctionStartedAt,
-			})
-			return {
-				completeSetCollateralAmount,
-				currentRetentionRate,
-				forkOutcome,
-				forkOwnSecurityPool,
-				hasForkActivity,
-				lastOraclePrice: lastSettlementTimestamp > 0n ? lastOraclePrice : undefined,
-				lastOracleSettlementTimestamp: lastSettlementTimestamp,
-				managerAddress,
-				marketDetails,
-				migratedRep,
-				parent,
-				questionOutcome: getReportingOutcomeKey(questionOutcome),
-				questionId: getQuestionIdHex(questionId),
-				securityMultiplier,
-				securityPoolAddress,
-				shareTokenSupply,
-				systemState,
-				totalRepDeposit,
-				totalSecurityBondAllowance,
-				truthAuctionAddress,
-				truthAuctionStartedAt,
-				universeHasForked: universeForkTime > 0n,
-				universeId,
-				hasLoadedVaults: vaultSummaries.hasLoadedVaults,
-				vaultCount: vaultSummaries.vaultCount,
-				vaults: vaultSummaries.vaults,
-			} satisfies ListedSecurityPool
-		}),
-	)
 	return {
 		pageIndex,
 		pageSize,
