+++ ui/ts/tests/securityPoolWorkflowSection/builders.ts
@@ -1,0 +1,337 @@
+import { zeroAddress, type Address } from 'viem'
+import { deriveHasForkActivity } from '../../lib/forkAuction.js'
+import type { AccountState } from '../../types/app.js'
+import type { ForkAuctionDetails, ListedSecurityPool, MarketDetails, OracleManagerDetails, SecurityPoolVaultSummary, SecurityVaultDetails } from '../../types/contracts.js'
+import type { ForkAuctionRouteContentProps, ReportingRouteContentProps, SecurityPoolWorkflowRouteContentProps, SecurityVaultRouteContentProps, TradingRouteContentProps } from '../../types/components.js'
+
+export function createAccountState(overrides: Partial<AccountState> = {}): AccountState {
+	return {
+		address: zeroAddress,
+		chainId: '0x1',
+		ethBalance: 0n,
+		wethBalance: 0n,
+		...overrides,
+	}
+}
+
+export function createTradingProps(overrides: Partial<TradingRouteContentProps> = {}): TradingRouteContentProps {
+	return {
+		accountState: createAccountState(),
+		loadingTradingForkUniverse: false,
+		loadingTradingDetails: false,
+		onCreateCompleteSet: () => undefined,
+		onMigrateShares: () => undefined,
+		onRedeemCompleteSet: () => undefined,
+		onRedeemShares: () => undefined,
+		onTradingFormChange: () => undefined,
+		repPerEthPrice: undefined,
+		repPerEthSource: undefined,
+		repPerEthSourceUrl: undefined,
+		selectedPool: undefined,
+		tradingActiveAction: undefined,
+		tradingDetails: undefined,
+		tradingError: undefined,
+		tradingForkUniverse: undefined,
+		tradingForm: {
+			completeSetAmount: '',
+			redeemAmount: '',
+			securityPoolAddress: '',
+			selectedShareOutcome: 'yes',
+			targetOutcomeIndexes: '',
+		},
+		tradingResult: undefined,
+		...overrides,
+	}
+}
+
+export function createReportingProps(overrides: Partial<ReportingRouteContentProps> = {}): ReportingRouteContentProps {
+	return {
+		accountState: createAccountState(),
+		loadingReportingDetails: false,
+		onLoadReporting: () => undefined,
+		onReportOutcome: () => undefined,
+		onReportingFormChange: () => undefined,
+		onWithdrawEscalation: (_outcome, _depositIndexes) => undefined,
+		reportingActiveAction: undefined,
+		reportingDetails: undefined,
+		reportingError: undefined,
+		reportingForm: {
+			reportAmount: '',
+			securityPoolAddress: '',
+			selectedOutcome: undefined,
+			selectedWithdrawDepositIndexesByOutcome: {
+				invalid: [],
+				yes: [],
+				no: [],
+			},
+		},
+		reportingResult: undefined,
+		...overrides,
+	}
+}
+
+export function createSecurityVaultProps(overrides: Partial<SecurityVaultRouteContentProps> = {}): SecurityVaultRouteContentProps {
+	return {
+		accountState: createAccountState(),
+		loadingSecurityVault: false,
+		onApproveRep: () => undefined,
+		onDepositRep: () => undefined,
+		onLoadSecurityVault: () => undefined,
+		onRedeemFees: () => undefined,
+		onRedeemRep: () => undefined,
+		onSetSecurityBondAllowance: () => undefined,
+		onSecurityVaultFormChange: () => undefined,
+		onWithdrawRep: () => undefined,
+		repPerEthPrice: undefined,
+		repPerEthSource: undefined,
+		repPerEthSourceUrl: undefined,
+		securityPoolVaults: undefined,
+		securityVaultActiveAction: undefined,
+		securityVaultDetails: undefined,
+		securityVaultError: undefined,
+		securityVaultForm: {
+			depositAmount: '',
+			repWithdrawAmount: '',
+			securityBondAllowanceAmount: '',
+			securityPoolAddress: '',
+			selectedVaultAddress: '',
+		},
+		securityVaultMissing: false,
+		securityVaultRepApproval: {
+			error: undefined,
+			loading: false,
+			value: 0n,
+		},
+		securityVaultRepBalance: undefined,
+		securityVaultResult: undefined,
+		selectedPoolSecurityMultiplier: undefined,
+		...overrides,
+	}
+}
+
+export function createSecurityVaultDetails(overrides: Partial<SecurityVaultDetails> = {}): SecurityVaultDetails {
+	return {
+		currentRetentionRate: 10n,
+		escalationEscrowedRep: 0n,
+		managerAddress: zeroAddress,
+		poolOwnershipDenominator: 1n,
+		repDepositShare: 5n * 10n ** 18n,
+		repToken: zeroAddress,
+		securityBondAllowance: 2n * 10n ** 18n,
+		securityPoolAddress: zeroAddress,
+		totalSecurityBondAllowance: 3n * 10n ** 18n,
+		unpaidEthFees: 1n * 10n ** 18n,
+		universeId: 1n,
+		vaultAddress: zeroAddress,
+		...overrides,
+	}
+}
+
+export function createOracleManagerDetails(overrides: Partial<OracleManagerDetails> = {}): OracleManagerDetails {
+	return {
+		callbackStateHash: undefined,
+		exactToken1Report: undefined,
+		isPriceValid: true,
+		lastPrice: 1n,
+		lastSettlementTimestamp: 1n,
+		managerAddress: zeroAddress,
+		openOracleAddress: zeroAddress,
+		pendingOperation: undefined,
+		pendingOperationSlotId: 0n,
+		pendingSettlementOperationIds: [],
+		pendingReportId: 0n,
+		priceRoundRemainingNotional: 1n,
+		priceValidUntilTimestamp: 1000n,
+		requestPriceEthCost: 1n,
+		token1: zeroAddress,
+		token2: zeroAddress,
+		...overrides,
+	}
+}
+
+export function createSecurityPoolVaultSummary(overrides: Partial<SecurityPoolVaultSummary> = {}): SecurityPoolVaultSummary {
+	return {
+		escalationEscrowedRep: 1n * 10n ** 18n,
+		repDepositShare: 5n * 10n ** 18n,
+		securityBondAllowance: 2n * 10n ** 18n,
+		unpaidEthFees: 1n * 10n ** 18n,
+		vaultAddress: zeroAddress,
+		...overrides,
+	}
+}
+
+export function createForkAuctionProps(overrides: Partial<ForkAuctionRouteContentProps> = {}): ForkAuctionRouteContentProps {
+	return {
+		accountState: createAccountState(),
+		forkAuctionActiveAction: undefined,
+		forkAuctionDetails: undefined,
+		forkAuctionError: undefined,
+		forkAuctionForm: {
+			claimBidIndex: '',
+			claimBidTick: '',
+			depositIndexes: '',
+			directForkQuestionId: '',
+			directForkUniverseId: '',
+			refundBidIndex: '',
+			refundTick: '',
+			repMigrationOutcomes: '',
+			securityPoolAddress: '',
+			selectedOutcome: 'yes',
+			settlementAddress: '',
+			submitBidAmount: '',
+			submitBidPrice: '',
+			vaultAddress: '',
+		},
+		forkAuctionResult: undefined,
+		loadingForkAuctionDetails: false,
+		onClaimAuctionProceeds: () => undefined,
+		onCreateChildUniverse: () => undefined,
+		onFinalizeTruthAuction: () => undefined,
+		onForkAuctionFormChange: () => undefined,
+		onForkUniverse: () => undefined,
+		onForkWithOwnEscalation: () => undefined,
+		onInitiateFork: () => undefined,
+		onLoadForkAuction: () => undefined,
+		onMigrateEscalationDeposits: (_outcome, _depositIndexes) => undefined,
+		onMigrateUnresolvedEscalation: _selectedChildOutcome => undefined,
+		onMigrateRepToZoltar: _outcomes => undefined,
+		onMigrateVault: () => undefined,
+		onRefundLosingBids: () => undefined,
+		onStartTruthAuction: () => undefined,
+		onSubmitBid: (_securityPoolAddressOverride?: Address) => undefined,
+		onWithdrawForkedEscalation: (_outcome, _parentDepositIndexes) => undefined,
+		...overrides,
+	}
+}
+
+export function createForkAuctionDetails(overrides: Partial<ForkAuctionDetails> = {}): ForkAuctionDetails {
+	const forkAuctionDetails: ForkAuctionDetails = {
+		auctionedSecurityBondAllowance: 0n,
+		auctionableRepAtFork: 0n,
+		claimingAvailable: false,
+		completeSetCollateralAmount: 0n,
+		currentTime: 3n,
+		forkOutcome: 'none',
+		forkOwnSecurityPool: false,
+		hasForkActivity: false,
+		marketDetails: createMarketDetails(),
+		migratedRep: 0n,
+		migrationEndsAt: undefined,
+		parentSecurityPoolAddress: zeroAddress,
+		questionOutcome: 'none',
+		securityPoolAddress: zeroAddress,
+		systemState: 'operational',
+		truthAuction: undefined,
+		truthAuctionAddress: zeroAddress,
+		truthAuctionStartedAt: 0n,
+		universeId: 1n,
+		...overrides,
+	}
+	return {
+		...forkAuctionDetails,
+		hasForkActivity: overrides.hasForkActivity ?? deriveHasForkActivity(forkAuctionDetails),
+	}
+}
+
+export function createMarketDetails(overrides: Partial<MarketDetails> = {}): MarketDetails {
+	return {
+		answerUnit: '',
+		createdAt: 1n,
+		description: 'Question description',
+		displayValueMax: 100n,
+		displayValueMin: 0n,
+		endTime: 2n,
+		exists: true,
+		marketType: 'binary',
+		numTicks: 2n,
+		outcomeLabels: ['Yes', 'No'],
+		questionId: '0x01',
+		startTime: 1n,
+		title: 'Will this resolve?',
+		...overrides,
+	}
+}
+
+export function createSelectedPool(overrides: Partial<ListedSecurityPool> = {}): ListedSecurityPool {
+	const selectedPool: ListedSecurityPool = {
+		completeSetCollateralAmount: 0n,
+		currentRetentionRate: 10n,
+		forkOutcome: 'none',
+		forkOwnSecurityPool: false,
+		hasForkActivity: false,
+		lastOraclePrice: undefined,
+		lastOracleSettlementTimestamp: 0n,
+		managerAddress: zeroAddress,
+		marketDetails: createMarketDetails(),
+		migratedRep: 0n,
+		parent: zeroAddress,
+		questionId: '0x01',
+		questionOutcome: 'none',
+		securityMultiplier: 2n,
+		securityPoolAddress: zeroAddress,
+		shareTokenSupply: 0n,
+		systemState: 'operational',
+		totalRepDeposit: 0n,
+		totalSecurityBondAllowance: 5n * 10n ** 18n,
+		truthAuctionAddress: zeroAddress,
+		truthAuctionStartedAt: 0n,
+		universeHasForked: false,
+		universeId: 1n,
+		vaultCount: 0n,
+		vaults: [],
+		...overrides,
+	}
+	return {
+		...selectedPool,
+		hasForkActivity: overrides.hasForkActivity ?? deriveHasForkActivity(selectedPool),
+	}
+}
+
+export function createSecurityPoolWorkflowProps(overrides: Partial<SecurityPoolWorkflowRouteContentProps> = {}): SecurityPoolWorkflowRouteContentProps {
+	return {
+		accountState: createAccountState(),
+		activeUniverseId: 1n,
+		checkedSecurityPoolAddress: undefined,
+		closeLiquidationModal: () => undefined,
+		forkAuction: createForkAuctionProps(),
+		liquidationAmount: '',
+		liquidationMaxAmount: undefined,
+		liquidationManagerAddress: undefined,
+		liquidationModalOpen: false,
+		liquidationSecurityPoolAddress: undefined,
+		liquidationTargetVault: '',
+		liquidationTimeoutMinutes: '5',
+		loadingPoolOracleManager: false,
+		loadingSecurityPools: false,
+		onExecutePendingPoolOperation: () => undefined,
+		onLiquidationAmountChange: () => undefined,
+		onLiquidationTimeoutMinutesChange: () => undefined,
+		onLoadPoolOracleManager: () => undefined,
+		onOpenLiquidationModal: () => undefined,
+		onQueueLiquidation: () => undefined,
+		onRefreshSelectedPoolData: () => undefined,
+		onRequestPoolPrice: () => undefined,
+		onSelectedPoolViewChange: () => undefined,
+		onSecurityPoolAddressChange: () => undefined,
+		onViewPendingReport: () => undefined,
+		poolOracleActiveAction: undefined,
+		poolOracleManagerDetails: undefined,
+		poolOracleManagerError: undefined,
+		poolPriceOracleResult: undefined,
+		repPerEthPrice: undefined,
+		repPerEthSource: undefined,
+		repPerEthSourceUrl: undefined,
+		reporting: createReportingProps(),
+		selectedPoolRefreshNonce: 0,
+		selectedPoolView: '',
+		securityPoolAddress: '',
+		securityPoolLiquidationError: undefined,
+		securityPoolOverviewActiveAction: undefined,
+		securityPoolOverviewError: undefined,
+		securityPoolOverviewResult: undefined,
+		securityPools: [],
+		securityVault: createSecurityVaultProps(),
+		trading: createTradingProps(),
+		...overrides,
+	}
+}
